Surface: SurfaceListener::onBufferDetached now returns buffer id
Previously returned slot, which wasn't partuclarly useful. As
evidenced by how it wasn't used anywhere.

Bug: 486182281
Flag: EXEMPT PURE_REFACTOR
Test: builds

+++ services/camera/libcameraservice/api1/client2/ZslProcessor.h
@@ -90,7 +90,7 @@ class ZslProcessor :
         virtual void onBufferReleased() override;
         virtual void onBuffersDiscarded(const std::vector<sp<GraphicBuffer>>& /* buffers */)
             override {}
-        virtual void onBufferDetached(int /* slot */) override {}
+        virtual void onBufferDetached(uint64_t /* bufferId */) override {}
         virtual bool needsReleaseNotify() override { return true; }
 
     private:
